+++ app/src/main/java/com/devscore/digital_pharmacy/business/datasource/cache/sales/FailureSalesOrderMedicineEntity.kt
@@ -27,7 +27,7 @@ data class FailureSalesOrderMedicineEntity (
 
     var unit : Int,
 
-    var quantity : Long,
+    var quantity : Float,
 
     var local_medicine : Int,
