+++ app/src/main/java/com/devscore/digital_pharmacy/business/datasource/cache/sales/SalesOrderMedicineEntity.kt
@@ -27,7 +27,7 @@ data class SalesOrderMedicineEntity (
 
     var unit : Int,
 
-    var quantity : Long,
+    var quantity : Float,
 
     var local_medicine : Int,
